+++ examples/components/navigation.js
@@ -2,7 +2,7 @@ import {Link} from 'react-router';
 import React, {PropTypes} from 'react';
 
 export function Navigation (props) {
-  const {example, examples, isRedux} = props;
+  const {example, examples} = props;
   const list = examples ? examples : [
     {key: 'back', name: 'Back', path: '/'},
     {key: example.key || 'loading', name: example.name}
@@ -18,7 +18,6 @@ export function Navigation (props) {
     }}>
       {list.map(({key, name, path}) => {
         if (!path) { return <span key={key}>{name}</span>; }
-        if (isRedux) { path += '?redux=true'; }
         if (key === 'back') { return <Link key={key} to={path}>{name}</Link>; }
 
         return (
@@ -35,35 +34,6 @@ export function Navigation (props) {
           </div>
         );
       })}
-      <a href={isRedux ? '?' : '?redux=true'} style={{
-        alignItems: 'center',
-        border: '2px solid rgb(0, 216, 255)',
-        borderRadius: '5vmin',
-        display: 'flex',
-        fontSize: '5vmin',
-        height: '8vmin',
-        textDecoration: 'none'
-      }}>
-        <div style={{
-          display: isRedux ? 'block' : 'none',
-          marginLeft: '1vmin',
-          textAlign: 'center',
-          width: '16vmin'
-        }}>{'redux'}</div>
-        <div style={{
-          backgroundColor: 'rgb(0, 216, 255)',
-          borderRadius: '4vmin',
-          height: '6vmin',
-          margin: '1vmin',
-          width: '6vmin'
-        }}/>
-        <div style={{
-          display: isRedux ? 'none' : 'block',
-          marginRight: '1vmin',
-          textAlign: 'center',
-          width: '16vmin'
-        }}>{'state'}</div>
-      </a>
     </div>
   );
 }
@@ -78,6 +48,5 @@ Navigation.propTypes = {
     key: PropTypes.string,
     name: PropTypes.string,
     path: PropTypes.string
-  })),
-  isRedux: PropTypes.bool
+  }))
 };