+++ examples/server.js
@@ -1,0 +1,3 @@
+require('babel-core/register');
+
+require('./' + process.env.npm_config_server);